Diff build process.
