+++ morf/setup.py
@@ -1,9 +1,6 @@
 import os
 
-from sklearn._build_utils import maybe_cythonize_extensions
-from sklearn._build_utils.deprecated_modules import (
-    _create_deprecated_modules_files
-)
+from morf._build_utils import maybe_cythonize_extensions
 
 
 def configuration(parent_package='', top_path=None):
@@ -14,23 +11,16 @@ def configuration(parent_package='', top_path=None):
     if os.name == 'posix':
         libraries.append('m')
 
-    _create_deprecated_modules_files()
-
     config = Configuration('morf', parent_package, top_path)
 
     # submodules which have their own setup.py
     config.add_subpackage('tree')
 
-    # add cython extension module for isotonic regression
-    config.add_extension('_isotonic',
-                         sources=['_isotonic.pyx'],
-                         include_dirs=[numpy.get_include()],
-                         libraries=libraries,
-                         )
-
     # add the test directory
-    config.add_subpackage('tests')
+    # config.add_subpackage('tests')
 
+    print('inside here....', top_path)
+    print(config)
     maybe_cythonize_extensions(top_path, config)
 
     return config
