--- morf/tree/__init__.py
@@ -1,3 +1,0 @@
-from ..forest import *
-
-__version__ = "0.1"
