+++ morf/tree/setup.py
@@ -9,8 +9,8 @@ def configuration(parent_package="", top_path=None):
     libraries = []
     if os.name == 'posix':
         libraries.append('m')
-    config.add_extension("split",
-                         sources=["split.pyx"],
+    config.add_extension("_split",
+                         sources=["_split.pyx"],
                          include_dirs=[numpy.get_include()],
                          libraries=libraries,
                          extra_compile_args=["-O3"],
